requirements.txt
asgiref==3.11.1
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.0
Django==6.0.2
django-cors-headers==4.9.0
django-jazzmin==3.0.2
django-rest-framework==0.1.0
djangorestframework==3.16.1
pillow==12.1.1
qrcode[pil]==7.4.2
reportlab==4.2.5
sqlparse==0.5.5
tzdata==2025.3
